Move logic from controller to separated model. Refactor code.
