+++ Service/CartAbandonmentIndexer.php
@@ -26,6 +26,11 @@ class CartAbandonmentIndexer implements \Buzzi\PublishCartAbandonment\Api\CartAb
     private $cartAbandonmentRepository;
 
     /**
+     * @var \Buzzi\PublishCartAbandonment\Model\ResourceModel\CartAbandonment
+     */
+    private $cartAbandonmentResource;
+
+    /**
      * @var \Magento\Eav\Model\Config
      */
     private $eavConfig;
@@ -34,17 +39,20 @@ class CartAbandonmentIndexer implements \Buzzi\PublishCartAbandonment\Api\CartAb
      * @param \Magento\Customer\Model\Visitor $visitorModel
      * @param \Magento\Reports\Model\ResourceModel\Quote\CollectionFactory $quoteCollectionFactory
      * @param \Buzzi\PublishCartAbandonment\Api\CartAbandonmentRepositoryInterface $cartAbandonmentRepository
-     * @param \Magento\Eav\Model\Config|null $eavConfig
+     * @param \Buzzi\PublishCartAbandonment\Model\ResourceModel\CartAbandonment $cartAbandonmentResource
+     * @param \Magento\Eav\Model\Config $eavConfig
      */
     public function __construct(
         \Magento\Customer\Model\Visitor $visitorModel,
         \Magento\Reports\Model\ResourceModel\Quote\CollectionFactory $quoteCollectionFactory,
         \Buzzi\PublishCartAbandonment\Api\CartAbandonmentRepositoryInterface $cartAbandonmentRepository,
+        \Buzzi\PublishCartAbandonment\Model\ResourceModel\CartAbandonment $cartAbandonmentResource,
         \Magento\Eav\Model\Config $eavConfig
     ) {
         $this->visitorModel = $visitorModel;
         $this->quoteCollectionFactory = $quoteCollectionFactory;
         $this->cartAbandonmentRepository = $cartAbandonmentRepository;
+        $this->cartAbandonmentResource = $cartAbandonmentResource;
         $this->eavConfig = $eavConfig;
     }
 
@@ -56,6 +64,7 @@ class CartAbandonmentIndexer implements \Buzzi\PublishCartAbandonment\Api\CartAb
      * @param bool $isResubmissionAllowed
      * @return void
      * @throws \Magento\Framework\Exception\CouldNotSaveException
+     * @throws \Magento\Framework\Exception\LocalizedException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
      */
     public function reindex(
@@ -70,13 +79,12 @@ class CartAbandonmentIndexer implements \Buzzi\PublishCartAbandonment\Api\CartAb
         $this->prepareFilters($quoteCollection, $quoteLastActionDays, $isRespectAcceptsMarketing, $storeId);
         $this->processQuoteLimit($quoteCollection, $quoteLimit);
 
-
         /** @var \Magento\Quote\Model\Quote $quote */
         foreach ($quoteCollection as $quote) {
             $cartAbandonment = $this->cartAbandonmentRepository->getByQuoteId($quote->getId(), true);
 
             $quoteFingerprint = $quote->getData('fingerprint');
-            $fingerprints = $this->cartAbandonmentRepository->getQuoteFingerprints($quote->getId());
+            $fingerprints = $this->cartAbandonmentResource->getQuoteFingerprints($quote->getId());
             if (!empty($fingerprints) && (!$isResubmissionAllowed || in_array($quoteFingerprint, $fingerprints))) {
                 continue;
             }
@@ -153,8 +161,7 @@ class CartAbandonmentIndexer implements \Buzzi\PublishCartAbandonment\Api\CartAb
     private function processQuoteLimit($quoteCollection, $quoteLimit)
     {
         if ($quoteLimit > 0) {
-            $quoteCollection->getSelect()
-                ->limit($quoteLimit);
+            $quoteCollection->getSelect()->limit($quoteLimit);
         }
     }
